requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
distro==1.9.0
Django==5.1.4
django-allauth==65.3.1
django-storages==1.14.4
djangorestframework==3.15.2
drf-yasg==1.21.8
filelock==3.16.1
fsspec==2024.12.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jiter==0.8.2
numpy==2.2.1
openai==1.58.1
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
pytest>=6.2.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests>=2.25.1
rsa==4.9
safetensors==0.5.0
six>=1.15.0
sniffio==1.3.1
sqlparse==0.5.3
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
confluent-kafka>=1.7.0
